refactor: 민감한정보 숨기기
