+++ app/build.gradle.kts
@@ -1,9 +1,14 @@
+import org.jetbrains.kotlin.konan.properties.Properties
+
 plugins {
     id("com.android.application")
     id("org.jetbrains.kotlin.android")
     id("de.mannodermaus.android-junit5") version "1.10.0.0"
     id("kotlin-kapt")
 }
+
+val properties = Properties()
+properties.load(project.rootProject.file("local.properties").inputStream())
 
 android {
     namespace = "woowacourse.shopping"
@@ -19,6 +24,24 @@ android {
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         testInstrumentationRunnerArguments["runnerBuilder"] = "de.mannodermaus.junit5.AndroidJUnit5Builder"
+
+        buildConfigField(
+            "String",
+            "BASE_URL",
+            properties.getProperty("BASE_URL"),
+        )
+
+        buildConfigField(
+            "String",
+            "USERNAME",
+            properties.getProperty("USERNAME"),
+        )
+
+        buildConfigField(
+            "String",
+            "PASSWORD",
+            properties.getProperty("PASSWORD"),
+        )
     }
 
     buildTypes {
