+++ app/src/main/java/woowacourse/shopping/view/detail/viewmodel/DetailViewModel.kt
@@ -9,10 +9,10 @@ import woowacourse.shopping.domain.model.RecentProduct
 import woowacourse.shopping.domain.repository.CartRepository
 import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.domain.repository.RecentProductRepository
-import woowacourse.shopping.util.Event
 import woowacourse.shopping.view.cart.listener.QuantityClickListener
 import woowacourse.shopping.view.cart.viewmodel.CartViewModel.Companion.DESCENDING_SORT_ORDER
 import woowacourse.shopping.view.detail.listener.DetailClickListener
+import woowacourse.shopping.view.event.Event
 import woowacourse.shopping.view.state.UiState
 
 class DetailViewModel(
